TDD - getPhenotype
