+++ src/main/scala/model/Bunny.scala
@@ -1,7 +1,4 @@
 package model
-import model.AlleleKind.AlleleKind
-import model.GeneKind.GeneKind
-
 import scala.util.Random
 
 object BunnyUtils {
@@ -22,24 +19,5 @@ case class Bunny(genotype: Genotype,
                  age: Int = 0,
                  isAlive: Boolean = true){
   def getTree(gens: Int) = ???
+  override def toString() = ???
 }
-
-case class Genotype(genes: List[Gene]) {
-  def + (gene :Gene) = ???
-}
-
-case class Allele(kind: AlleleKind,
-                  isDominant: Option[Boolean] = Option.empty,
-                  isMutated: Option[Boolean] = Option.empty)
-
-case class Gene(kind: GeneKind,
-                momAllele: Allele,
-                dadAllele: Allele,
-                letters: Option[String] = Option.empty) {
-  def getPhenotype(): Phenotype = ???
-
-  if (!(momAllele.kind == kind.mutated || momAllele.kind == kind.base) && (dadAllele.kind == kind.mutated || dadAllele.kind == kind.base))
-    throw new IllegalAlleleException("Gene initialization: one of the Alleles (momAllele or dadAllele) has a kind which is not suitable with the kind of the Gene!")
-}
-
-case class Phenotype(traits: Seq[AlleleKind])
